Publish gRPC endpoint metadata through compiler lifecycle
