+++ compiler-plugin/src/main/java/io/ballerina/stdlib/grpc/plugin/endpointyaml/generator/Endpoint.java
@@ -19,7 +19,7 @@
 package io.ballerina.stdlib.grpc.plugin.endpointyaml.generator;
 
 /**
- * Represents the metadata of a single gRPC endpoint exported to {@code endpoints.yaml}.
+ * Represents the metadata of a single gRPC endpoint.
  */
 public class Endpoint {
     private final String name;
